feat-eng/ex4: fix NameError exception
This fixes a NameError exception due to the 'selected' variable not
being properly exported from L1Regularization's 'check' method in
learntools/feature_engineering/ex4.py. Instead of declaring 'selected'
as global before assigning it, this augments the cell which uses the
'selected' variable with the code from the 'check' method. This way it's
more obvious for students where the variable originates from rather than
magically assigning it behind the scenes.

This fixes #229.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>