fixed int array error in learntools SQL ex5 q3
The bug was in lines 100 and 101, where the code was trying to call int() on a single-element integer list, which returned `TypeError: int() argument must be a string, a bytes-like object or a real number, not 'IntegerArray'`. To fix this, all I did was replace the `int()` function with accessing the first (and only) element with `[0]` for both lines.