Sort the results before comparison
The SQL statements don't contain an `ORDER BY` clause, so the ordering
can't be an intentional check for this exercise. This exercise happens
to work most of the time because BigQuery tends to return results in the
same order, but sometimes it doesn't. Learn users shouldn't have to deal
with this non-determinism in the check.

http://b/415948668